requirements.txt
pyrebase4
requests 
protobuf
asn1crypto
certvalidator
chily
sgx-ias-structs
